Modify option parsing to allow "-" as an argument.
We need to overhaul this eventually; the option parser isn't given
enough information to do proper verification.  GLib does this
fairly well, and boost's program_options library does so as well;
we probably don't need anything as complex as the latter, but
might want to look at their ideas sometime.

git-svn-id: https://gforge.sci.utah.edu/svn/basics/trunk@54 5b8196f7-c1c0-4be6-8b13-4feed349168d