LargeFile simplifications.
Add a LargeFileFD class which can both read from a file
descriptor, as well as serving as a base for LargeFileMMap, and
(in the future) LargeFileAIO.

Switch to shared_ptrs instead of weak_ptrs.  Weak references make
a bit of sense for mmap, but are a real PITA when you need
explicit buffers, as one does in the FD and AIO cases.

Be more proactive about keeping the memory we map slim.  If we're
opening a file readonly, for example, just map the parts of the
file that exist.  The operating system will go nuts about our use
of virtual memory if we don't do this.

git-svn-id: https://gforge.sci.utah.edu/svn/basics/trunk@172 5b8196f7-c1c0-4be6-8b13-4feed349168d